refacto: storybook directory
