+++ ui/tools/storybook/stories/testing/MockInteractiveImageSegmenter.ts
