+++ ui/tools/storybook/stories/testing/datasets/index.ts
@@ -1,0 +1,2 @@
+export * from "./mockImageDatasetItem";
+export * from "./mockVideoDatasetItem";
