+++ ui/tools/storybook/stories/testing/index.ts
@@ -1,0 +1,3 @@
+export * from "./mockHandleSaveItem";
+export * from "./MockInteractiveImageSegmenter";
+export * from "./shared";
