+++ ui/tools/storybook/stories/testing/mockHandleSaveItem.ts
@@ -1,0 +1,9 @@
+/*-------------------------------------
+Copyright: CEA-LIST/DIASI/SIALV/LVA
+Author : pixano@cea.fr
+License: CECILL-C
+-------------------------------------*/
+
+import type { DatasetItem } from "@pixano/core";
+
+export const mockHandleSaveItem = (item: DatasetItem) => Promise.resolve();
